requirements.txt
openai
langchain==0.3.6
langchain-core==0.3.15
langchain-openai
langchain-community
python-dotenv
streamlit
